# frontend/src/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/polls": {
      "post": {
        "operationId": "PollController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePollDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } }
      },
      "get": {
        "operationId": "PollController_findAll",
        "parameters": [],
        "responses": { "200": { "description": "" } }
      }
    },
    "/polls/{pollId}": {
      "get": {
        "operationId": "PollController_findOne",
        "parameters": [
          {
            "name": "pollId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "" } }
      },
      "patch": {
        "operationId": "PollController_update",
        "parameters": [
          {
            "name": "pollId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePollDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } }
      },
      "delete": {
        "operationId": "PollController_remove",
        "parameters": [
          {
            "name": "pollId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "" } }
      }
    },
    "/polls/byCode/{code}": {
      "get": {
        "operationId": "PollController_findByCode",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } }
      }
    }
  },
  "info": {
    "title": "Solid Polls",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    { "url": "http://localhost:3000", "description": "Testing" },
    { "url": "https://solidpolls.de/api", "description": "Production" }
  ],
  "components": {
    "schemas": {
      "CreatePollDto": {
        "type": "object",
        "properties": { "title": { "type": "string" } },
        "required": ["title"]
      },
      "UpdatePollDto": {
        "type": "object",
        "properties": { "title": { "type": "string" } }
      }
    }
  }
}